fix(ci): scope projection and recovery checks
