+++ src/gateway/session-display-projection.test.ts
@@ -1,0 +1,51 @@
+import { describe, expect, test } from "vitest";
+import { projectSessionDisplayMessage } from "./session-display-projection.js";
+
+const SESSION_LAST_MESSAGE_PREVIEW_MAX_CHARS = 240;
+
+describe("projectSessionDisplayMessage", () => {
+  test("keeps visible user and assistant text while excluding non-display rows", () => {
+    const messages = [
+      { role: "user", content: "Initial request" },
+      { role: "assistant", content: "Visible final answer" },
+      { role: "toolResult", content: [{ type: "text", text: "tool output" }] },
+      { role: "system", content: "system event" },
+      {
+        role: "assistant",
+        content: [
+          { type: "thinking", thinking: "private thought" },
+          { type: "reasoning", text: "reasoning summary" },
+        ],
+      },
+      { role: "assistant", content: "NO_REPLY" },
+      {
+        role: "assistant",
+        content: [{ type: "text", text: "" }],
+        openclawDelivery: { replyToCurrent: true },
+      },
+    ];
+
+    expect(messages.map(projectSessionDisplayMessage)).toEqual([
+      { role: "user", text: "Initial request" },
+      { role: "assistant", text: "Visible final answer" },
+      null,
+      null,
+      null,
+      null,
+      null,
+    ]);
+  });
+
+  test("bounds previews without splitting surrogate pairs", () => {
+    const longReply = `${"a".repeat(SESSION_LAST_MESSAGE_PREVIEW_MAX_CHARS - 2)}😊tail`;
+    const preview = projectSessionDisplayMessage({ role: "assistant", content: longReply });
+
+    expect(preview?.text).toHaveLength(SESSION_LAST_MESSAGE_PREVIEW_MAX_CHARS);
+    expect(preview?.text).toBe(`${"a".repeat(SESSION_LAST_MESSAGE_PREVIEW_MAX_CHARS - 3)}...`);
+  });
+
+  test("preserves quoted directive examples", () => {
+    const quoted = "Use `[[reply_to_current]]` literally.";
+    expect(projectSessionDisplayMessage({ role: "assistant", content: quoted })?.text).toBe(quoted);
+  });
+});
